Docs Updated for changes, Change for java examples
