+++ examples/src/main/python/kmeans.py
@@ -18,7 +18,7 @@
 """
 The K-means algorithm written from scratch against PySpark. In practice,
 one may prefer to use the KMeans algorithm in MLlib, as shown in
-python/examples/mllib/kmeans.py.
+examples/src/main/python/mllib/kmeans.py.
 
 This example requires NumPy (http://www.numpy.org/).
 """
