+++ examples/src/main/scala/org/apache/spark/examples/streaming/NetworkWordCount.scala
@@ -32,7 +32,7 @@ import org.apache.spark.storage.StorageLevel
  * To run this on your local machine, you need to first run a Netcat server
  *    `$ nc -lk 9999`
  * and then run the example
- *    `$ ./bin/run-example org.apache.spark.streaming.examples.NetworkWordCount local[2] localhost 9999`
+ *    `$ ./bin/run-example org.apache.spark.examples.streaming.NetworkWordCount local[2] localhost 9999`
  */
 // scalastyle:on
 object NetworkWordCount {
